fix linux build
git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@287 fea1f14b-4e95-431d-9825-d582385f3f73

+++ empyrean/src/base/Connection.cpp
@@ -22,7 +22,7 @@ namespace pyr {
     }
     
     Connection::~Connection() {
-        for (unsigned i = 0; i < _unhandledPackets.size(); ++i) {
+        for (size_t i = 0; i < _unhandledPackets.size(); ++i) {
             delete _unhandledPackets[i];
         }
         for (HandlerMapItr i = _handlers.begin(); i != _handlers.end(); ++i) {
